requirements.txt
reflex>=0.4.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
akshare>=1.12.0
plotly>=5.18.0
zhipuai>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
loguru>=0.7.0
